Moved registry to coordinator package
Since it is coordinator specific functionality. It isn't part of the protocol.
