+++ simulator/src/main/java/com/hazelcast/simulator/coordinator/ConfigFileTemplate.java
@@ -16,8 +16,8 @@
 
 package com.hazelcast.simulator.coordinator;
 
-import com.hazelcast.simulator.protocol.registry.AgentData;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.AgentData;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
 import com.hazelcast.simulator.utils.TagUtils;
 import freemarker.template.Configuration;
 import freemarker.template.SimpleScalar;
