+++ simulator/src/main/java/com/hazelcast/simulator/coordinator/CoordinatorCli.java
@@ -21,9 +21,9 @@ import com.hazelcast.simulator.common.TestPhase;
 import com.hazelcast.simulator.common.WorkerType;
 import com.hazelcast.simulator.coordinator.tasks.DownloadTask;
 import com.hazelcast.simulator.coordinator.tasks.ArtifactCleanTask;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
-import com.hazelcast.simulator.protocol.registry.TargetType;
-import com.hazelcast.simulator.protocol.registry.WorkerQuery;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.TargetType;
+import com.hazelcast.simulator.coordinator.registry.WorkerQuery;
 import com.hazelcast.simulator.utils.CommandLineExitException;
 import joptsimple.OptionParser;
 import joptsimple.OptionSet;
