+++ simulator/src/main/java/com/hazelcast/simulator/coordinator/CoordinatorRemoteCli.java
@@ -31,8 +31,8 @@ import com.hazelcast.simulator.protocol.operation.RcWorkerKillOperation;
 import com.hazelcast.simulator.protocol.operation.RcWorkerScriptOperation;
 import com.hazelcast.simulator.protocol.operation.RcWorkerStartOperation;
 import com.hazelcast.simulator.protocol.operation.SimulatorOperation;
-import com.hazelcast.simulator.protocol.registry.TargetType;
-import com.hazelcast.simulator.protocol.registry.WorkerQuery;
+import com.hazelcast.simulator.coordinator.registry.TargetType;
+import com.hazelcast.simulator.coordinator.registry.WorkerQuery;
 import com.hazelcast.simulator.utils.CommandLineExitException;
 import com.hazelcast.simulator.utils.FileUtils;
 import com.hazelcast.simulator.utils.TagUtils;
