+++ simulator/src/main/java/com/hazelcast/simulator/coordinator/TestCaseRunner.java
@@ -22,10 +22,10 @@ import com.hazelcast.simulator.protocol.operation.CreateTestOperation;
 import com.hazelcast.simulator.protocol.operation.StartTestOperation;
 import com.hazelcast.simulator.protocol.operation.StartTestPhaseOperation;
 import com.hazelcast.simulator.protocol.operation.StopTestOperation;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
-import com.hazelcast.simulator.protocol.registry.TargetType;
-import com.hazelcast.simulator.protocol.registry.TestData;
-import com.hazelcast.simulator.protocol.registry.WorkerData;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.TargetType;
+import com.hazelcast.simulator.coordinator.registry.TestData;
+import com.hazelcast.simulator.coordinator.registry.WorkerData;
 import org.apache.log4j.Logger;
 
 import java.util.ArrayList;
@@ -44,8 +44,8 @@ import static com.hazelcast.simulator.common.TestPhase.LOCAL_TEARDOWN;
 import static com.hazelcast.simulator.common.TestPhase.LOCAL_VERIFY;
 import static com.hazelcast.simulator.common.TestPhase.RUN;
 import static com.hazelcast.simulator.common.TestPhase.SETUP;
-import static com.hazelcast.simulator.protocol.registry.TestData.CompletedStatus.FAILED;
-import static com.hazelcast.simulator.protocol.registry.TestData.CompletedStatus.SUCCESS;
+import static com.hazelcast.simulator.coordinator.registry.TestData.CompletedStatus.FAILED;
+import static com.hazelcast.simulator.coordinator.registry.TestData.CompletedStatus.SUCCESS;
 import static com.hazelcast.simulator.utils.CommonUtils.await;
 import static com.hazelcast.simulator.utils.CommonUtils.getElapsedSeconds;
 import static com.hazelcast.simulator.utils.CommonUtils.rethrow;
