+++ simulator/src/main/java/com/hazelcast/simulator/coordinator/registry/TestData.java
@@ -13,16 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.simulator.protocol.registry;
+package com.hazelcast.simulator.coordinator.registry;
 
 import com.hazelcast.simulator.common.TestCase;
 import com.hazelcast.simulator.common.TestPhase;
 import com.hazelcast.simulator.coordinator.TestSuite;
 import com.hazelcast.simulator.protocol.core.SimulatorAddress;
 
-import static com.hazelcast.simulator.protocol.registry.TestData.CompletedStatus.FAILED;
-import static com.hazelcast.simulator.protocol.registry.TestData.CompletedStatus.IN_PROGRESS;
-import static com.hazelcast.simulator.protocol.registry.TestData.CompletedStatus.SUCCESS;
+import static com.hazelcast.simulator.coordinator.registry.TestData.CompletedStatus.FAILED;
+import static com.hazelcast.simulator.coordinator.registry.TestData.CompletedStatus.IN_PROGRESS;
+import static com.hazelcast.simulator.coordinator.registry.TestData.CompletedStatus.SUCCESS;
 
 public class TestData {
