+++ simulator/src/main/java/com/hazelcast/simulator/protocol/operation/RcWorkerKillOperation.java
@@ -15,7 +15,7 @@
  */
 package com.hazelcast.simulator.protocol.operation;
 
-import com.hazelcast.simulator.protocol.registry.WorkerQuery;
+import com.hazelcast.simulator.coordinator.registry.WorkerQuery;
 
 public class RcWorkerKillOperation implements SimulatorOperation {
     private final String command;
