+++ simulator/src/test/java/com/hazelcast/simulator/common/AgentsFileTest.java
@@ -1,7 +1,7 @@
 package com.hazelcast.simulator.common;
 
-import com.hazelcast.simulator.protocol.registry.AgentData;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.AgentData;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
 import com.hazelcast.simulator.utils.CommandLineExitException;
 import org.junit.After;
 import org.junit.Before;
