+++ simulator/src/test/java/com/hazelcast/simulator/coordinator/FailureCollectorTest.java
@@ -6,8 +6,8 @@ import com.hazelcast.simulator.common.WorkerType;
 import com.hazelcast.simulator.protocol.core.AddressLevel;
 import com.hazelcast.simulator.protocol.core.SimulatorAddress;
 import com.hazelcast.simulator.protocol.operation.FailureOperation;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
-import com.hazelcast.simulator.protocol.registry.WorkerData;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.WorkerData;
 import com.hazelcast.simulator.utils.TestUtils;
 import org.junit.After;
 import org.junit.Before;
