+++ simulator/src/test/java/com/hazelcast/simulator/coordinator/tasks/RunTestSuiteTaskTest.java
@@ -21,10 +21,10 @@ import com.hazelcast.simulator.protocol.operation.SimulatorOperation;
 import com.hazelcast.simulator.protocol.operation.StartTestOperation;
 import com.hazelcast.simulator.protocol.operation.StartTestPhaseOperation;
 import com.hazelcast.simulator.protocol.operation.StopTestOperation;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
-import com.hazelcast.simulator.protocol.registry.TargetType;
-import com.hazelcast.simulator.protocol.registry.TestData;
-import com.hazelcast.simulator.protocol.registry.WorkerQuery;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.TargetType;
+import com.hazelcast.simulator.coordinator.registry.TestData;
+import com.hazelcast.simulator.coordinator.registry.WorkerQuery;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
