+++ simulator/src/test/java/com/hazelcast/simulator/protocol/processors/TestOperationProcessorTest.java
@@ -14,7 +14,7 @@ import com.hazelcast.simulator.protocol.operation.SimulatorOperation;
 import com.hazelcast.simulator.protocol.operation.StartTestOperation;
 import com.hazelcast.simulator.protocol.operation.StartTestPhaseOperation;
 import com.hazelcast.simulator.protocol.operation.StopTestOperation;
-import com.hazelcast.simulator.protocol.registry.TargetType;
+import com.hazelcast.simulator.coordinator.registry.TargetType;
 import com.hazelcast.simulator.tests.FailingTest;
 import com.hazelcast.simulator.tests.SuccessTest;
 import com.hazelcast.simulator.worker.Worker;
