+++ simulator/src/test/java/com/hazelcast/simulator/utils/HazelcastUtilsTest.java
@@ -4,8 +4,8 @@ import com.hazelcast.core.Cluster;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.Member;
 import com.hazelcast.simulator.common.WorkerType;
-import com.hazelcast.simulator.protocol.registry.AgentData;
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.AgentData;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Test;
