+++ simulator/src/test/java/com/hazelcast/simulator/utils/SimulatorUtilsTest.java
@@ -1,6 +1,6 @@
 package com.hazelcast.simulator.utils;
 
-import com.hazelcast.simulator.protocol.registry.ComponentRegistry;
+import com.hazelcast.simulator.coordinator.registry.ComponentRegistry;
 import joptsimple.OptionSet;
 import joptsimple.OptionSpec;
 import org.junit.After;
